pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-auth>=2.40.3",
    "pyjwt>=2.10.1",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
